Update the CreateAccount method to a valid flow
Calling ResetPassword on an account that is unverified will raise the AccountCreatedEvent instead of ResetPasswordRequested, which does not actually allow the receiver to reset their password with that verification key.

VerifyEmailFrom Key sets the IsAccountVerified flag to true and adds a login date to the user, ensuring that the user can get a ResetPasswordRequested event.

Also, need to set the account to the UserAccountCreatedByAdmin event so that the email tokenizer doesn't fail.