AdminUserAccountService.CreateAccount cleanup
SetConfirmedEmail sets the IsAccountVerified flag to true, but it doesn't do it in the existing account object, which is why ResetPassword was failing.  Removed the VerifyEmailFromKey, readded SetConfirmedEmail, and added a call to GetByID to get the updated account.

This change meant that we no longer needed the user's password or the AccountCreated event, so cleaned that.

ResetPassword call is also unnecessary, as we can call SetVerificationKey directly and not have to generate our own.  Saves a lot of calls and a lot of checks that are redundant, tedious, bug-laden and unnecessary for what we're doing.

Moved the Tracing success message to the other side of our true Update call.