Remove IApplicationBuilder as a dependency
Removed IApplicationBuilder as a dependency of AuthCentralAppInfo.  IApplicationBuilder was used to get the HttpContext.  This can be obtained from IHttpContextAccessor, which is already available in IServiceCollection.  This allows MembershipRebootConfiguration objects to be created on the fly by the DI container, instead of being forced to create a singleton at the beginning.  This was necessary so as to avoid creating another singleton for the AdminConfig.  Also, if the general advice for these scenarios is to have multiple configurations, creating a singleton for each starts going down a dark rabbit hole.

Say no to singleton dependencies.